backend/requirements.txt
astroid==2.3.3
autopep8==1.4.4
beautifulsoup4==4.8.2
blinker==1.4
celery==4.4.2
Click==7.0
Flask==1.1.1
flask-apispec==0.8.7
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.1
flake8==3.8.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.5.0
mccabe==0.6.1
Pillow==7.0.0
psycopg2-binary==2.8.4
pycodestyle==2.6.0
# pylint==2.4.4
pymongo==3.9.0
redis==3.3.11
requests==2.23.0
six==1.13.0
SQLAlchemy==1.3.12
urllib3==1.25.8
# uwsgi == 2.0.17
Werkzeug==0.16.0
wrapt==1.11.2
